move tests to unittest & optimize module exports
