+++ angr_platforms/bf/__init__.py
@@ -1,1 +1,5 @@
-from . import arch_bf, load_bf, lift_bf, engine_bf, simos_bf
+from .arch_bf import ArchBF
+from .load_bf import BF
+from .lift_bf import LifterBF
+from .engine_bf import UberEngineWithBF
+from .simos_bf import SimBF, SimBFSyscall
